chore: extract broadcast-replay suite into src/suites/broadcast-replay.ts
Part of REAL-1037.
